Refactor: extract ErrorFixerController from VideoLoaderController
